requirements.txt
pip
pytz
Django
asgiref
sqlparse
setuptools
djangorestframework